Abstraction of logic and few minor fixes
