+++ ui/public/locales/en.json
@@ -576,7 +576,7 @@
 "label.create.template": "Create Template",
 "label.create.tier.aclid.description": "The ACL associated with the Network Tier.",
 "label.create.tier.externalid.description": "ID of the Network in an external system.",
-"label.create.tier.gateway.description": "Gateway in range of VPC CIDR ({value})",
+"label.create.tier.gateway.description": "Gateway IP must be within VPC CIDR ({value})",
 "label.create.tier.name.description": "A unique name for the Network Tier.",
 "label.create.tier.netmask.description": "The Network Tier's netmask. For example {value}",
 "label.create.tier.networkofferingid.description": "The Network offering for the Network Tier.",
